Improve container-engine/docker role
The current implementation of this role is verified only on Ubuntu
18.04 LTS. This change pretends to increase its coverage including
other distributions.
