--- roles/container-engine/docker/molecule/default/converge.yml
@@ -1,7 +1,0 @@
----
-- name: Converge
-  hosts: all
-  become: true
-  roles:
-    - role: kubespray-defaults
-    - role: docker
