--- roles/container-engine/docker/molecule/default/molecule.yml
@@ -1,28 +1,0 @@
----
-driver:
-  name: vagrant
-  provider:
-    name: libvirt
-lint: |
-  set -e
-  yamllint -c ../../../.yamllint .
-platforms:
-  - name: kubespray-docker
-    box: generic/ubuntu1804
-    cpus: 2
-    memory: 1024
-provisioner:
-  name: ansible
-  env:
-    ANSIBLE_ROLES_PATH: ../../../../
-  config_options:
-    defaults:
-      callback_whitelist: profile_tasks
-  lint:
-    name: ansible-lint
-    options:
-      c: ../../../.ansible-lint
-verifier:
-  name: testinfra
-  lint:
-    name: flake8
