+++ roles/container-engine/docker/molecule/resources/converge.yml
@@ -1,0 +1,16 @@
+---
+# SPDX-license-identifier: Apache-2.0
+##############################################################################
+# Copyright (c) 2020
+# All rights reserved. This program and the accompanying materials
+# are made available under the terms of the Apache License, Version 2.0
+# which accompanies this distribution, and is available at
+# http://www.apache.org/licenses/LICENSE-2.0
+##############################################################################
+
+- name: Converge
+  hosts: all
+  become: true
+  roles:
+    - role: kubespray-defaults
+    - role: docker
