+++ roles/container-engine/docker/molecule/ubuntu/molecule.yml
@@ -1,0 +1,53 @@
+---
+# SPDX-license-identifier: Apache-2.0
+##############################################################################
+# Copyright (c) 2020
+# All rights reserved. This program and the accompanying materials
+# are made available under the terms of the Apache License, Version 2.0
+# which accompanies this distribution, and is available at
+# http://www.apache.org/licenses/LICENSE-2.0
+##############################################################################
+
+driver:
+  name: vagrant
+  provider:
+    name: libvirt
+lint: |
+  set -e
+  yamllint -c ../../../.yamllint .
+platforms:
+  - name: xenial
+    box: generic/ubuntu1604
+    cpus: 1
+    memory: 1024
+  - name: bionic
+    box: generic/ubuntu1804
+    cpus: 1
+    memory: 1024
+  - name: focal
+    box: generic/ubuntu2004
+    cpus: 1
+    memory: 1024
+provisioner:
+  name: ansible
+  playbooks:
+    converge: ../resources/converge.yml
+    prepare: ../resources/prepare.yml
+  env:
+    ANSIBLE_ROLES_PATH: ../../../../
+  config_options:
+    defaults:
+      callback_whitelist: profile_tasks
+  lint:
+    name: ansible-lint
+    options:
+      c: ../../../.ansible-lint
+  inventory:
+    host_vars:
+      xenial:
+        docker_dns_servers_strict: false
+verifier:
+  name: testinfra
+  directory: ../resources/tests/
+  lint:
+    name: flake8
