+++ roles/container-engine/docker/tasks/pre-upgrade.yml
@@ -15,6 +15,14 @@
     - ansible_os_family == 'Debian'
     - (docker_versioned_pkg[docker_version | string] is search('docker-ce'))
 
+- name: Ensure APT transport is installed. | Debian
+  apt:
+    name: apt-transport-https
+    state: present
+  when:
+    - ansible_os_family == 'Debian'
+    - (docker_versioned_pkg[docker_version | string] is search('docker-ce'))
+
 - name: Ensure old versions of Docker are not installed. | RedHat
   yum:
     name: '{{ docker_remove_packages_yum }}'
